semester_4/OOP/lab3/pyproject.toml
[project]
name = "lab3"
version = "0.1.0"
description = "Image Editor"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "kivy==2.2.0",
    "pillow==10.0.1",
]

[project.scripts]
imagedit = "main:main"
